backend/requirements.txt
langchain==0.0.350
langchain-community==0.0.13
langchain-core==0.1.10
langchain-google-genai==0.0.5
pypdf==3.17.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
chromadb==0.4.22
tiktoken==0.5.2
pydantic==1.10.13 